requirements.txt
streamlit
qrcode[pil]
pillow
opencv-python-headless
numpy